requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chromedriver==2.24.1
chromedriver-autoinstaller==0.6.4
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.116.1
fdb==2.0.3
firebird-base==2.0.0
Flask==3.1.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==2.2.4
oauthlib==3.2.2
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2025.1
pywin32-ctypes==0.2.3
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
selenium==4.29.0
setuptools==77.0.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.47.1
tkcalendar==1.6.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.35.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
XlsxWriter==3.2.3
zeep==4.3.1
